% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_presentacion.R
\name{crear_presentacion}
\alias{crear_presentacion}
\title{Función para crear y abrir un template de presentaciones para la DNMyE}
\usage{
crear_presentacion(
  template = "skeleton.Rmd",
  save_as = template,
  data = list(),
  ignore = FALSE,
  open = TRUE,
  package = "comunicacion"
)
}
\arguments{
\item{template}{Ruta donde se ubica el template de referencia}

\item{save_as}{Ruta donde será guardado el nuevo informe creado}

\item{data}{Una lista de 'data' que se le pasa al template}

\item{ignore}{Si es TRUE, el nuevo archivo creado se incluye en \code{.Rbuildignore}. Por default, esta opción figura en FALSE}

\item{open}{Si es TRUE se abre el archivo al mismo tiempo que se crea, para poder editarlo.}

\item{package}{Nombre del paquete donde se encuentra el template.}
}
\value{
un vector lógico que indica si el archivo fue modificado
}
\description{
Función para crear y abrir un template de presentaciones para la DNMyE
}
\examples{
\dontrun{
  # Nota: Al correr la siguiente sentencia el archivo `presentacion.Rmd` se
  crear_informe(save_as = "presentacion.Rmd",
               open = TRUE)
}
}
