% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnmye_colores.R
\name{dnmye_colores}
\alias{dnmye_colores}
\title{Funcion para extraer la paleta institucional de la dnmye como colores hex}
\usage{
dnmye_colores(..., nombre = FALSE)
}
\arguments{
\item{...}{Nombre o número del color. Las posibilidades son \emph{rosa}, \emph{naranja}, \emph{amarillo}, \emph{pera}, \emph{azul verde}, \emph{purpura}, \emph{cian}, \emph{gris oscuro}, \emph{gris medio} y \emph{gris claro} (o del 1 al 10).}

\item{nombre}{Si es TRUE devuelve el nombre del color al que corresponde el código hex}
}
\value{
Paleta de colores institucionales para la Republica de Argentina
}
\description{
Funcion para extraer la paleta institucional de la dnmye como colores hex
}
\examples{
### Obtener el código hex del color institucional celeste
dnmye_colores()

dnmye_colores("rosa")
}
