% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnmye_paletas.R
\name{dnmye_paletas}
\alias{dnmye_paletas}
\title{Funcion para extraer colores de la dnmye como colores hex}
\usage{
dnmye_paletas(palette = "c10_todos", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Nombre de la paleta de colores en dnmye_paletas. Se puede optar por \emph{c2_continuo}, \emph{c2_contraste} o \emph{c10_todos}}

\item{reverse}{\emph{TRUE} indica que la peleta debe ser invertida}

\item{...}{Argumentos adicionales para definir con colorRampPalette()}
}
\value{
Paleta de colores institucionales para la Republica de Argentina
}
\description{
Funcion para extraer colores de la dnmye como colores hex
}
