% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_font.R
\docType{data}
\name{font_encode_regular}
\alias{font_encode_regular}
\alias{font_encode_bold}
\title{Encode font name R variable aliases}
\format{
length 1 character vector

length 1 character vector
}
\usage{
font_encode_regular

font_encode_bold
}
\description{
\code{font_encode_regular} == "\code{EncodeSansNormal-400-Regular}"

\code{font_encode_bold} == "\code{EncodeSansNormal-700-Bold}"
}
\keyword{datasets}
