% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_theme_dnmye.R
\name{gt_theme_dnmye}
\alias{gt_theme_dnmye}
\title{Función para estilizar tablas gt con theme propio}
\usage{
gt_theme_dnmye(gt_table, var_total = NA)
}
\arguments{
\item{gt_table}{Tabla gt}

\item{var_total}{Nombre de la variable que contiene la categoría "Total"}
}
\description{
Función para estilizar tablas gt con theme propio
}
