% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbl_int.R
\name{lbl_int}
\alias{lbl_int}
\title{Función para formatear números enteros como texto}
\usage{
lbl_int(x)
}
\arguments{
\item{x}{Número a formatear}
}
\description{
Función para formatear números enteros como texto
}
