% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdautotabs.R
\name{rmdautotabs}
\alias{rmdautotabs}
\title{Generacion automatica de tabs en Rmd}
\usage{
rmdautotabs(
  input_data,
  panel_name,
  ...,
  tabset_title = "",
  tabset_props = ".tabset-fade .tabset-pills",
  chunk_props = list(echo = FALSE, fig.align = "center"),
  is_output_distill = TRUE
)
}
\arguments{
\item{input_data}{Ungrouped tibble with at least 2 columns, one for the title of the tabs and another with the output to be displayed.}

\item{panel_name}{column with the ID variable.}

\item{...}{nested columns that contain outputs to display}

\item{tabset_title}{string title of the .tabset}

\item{tabset_props}{string defining .tabset properties. Only works with is_output_distill = F}

\item{chunk_props}{named list with additional parameters that correspond to all those available in rmarkdown chunks (fig.align, fig.width, ...).}

\item{is_output_distill}{boolean. is output a distill article.}
}
\description{
Permite generar automáticamente el código necesario para agrupar múltiples
fragmentos de Rmarkdown en pestañas. Concatena todos los fragmentos en una
cadena que luego se puede knitear y renderizar. Tomado de {sknifedatar}
}
