% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_dnmye.R
\name{scale_color_dnmye}
\alias{scale_color_dnmye}
\title{Constructor de escala de colores de contorno (color) para la dnmye}
\usage{
scale_color_dnmye(
  palette = "cualitativa",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Nombre de la paleta de colores en dnmye_paleta}

\item{discrete}{TRUE indica que la variable es discreta}

\item{reverse}{\emph{TRUE} indica que la peleta debe ser invertida}

\item{...}{Argumentos adicionales para definir en discrete_scale() o
scale_color_gradientn(), respectivamente, cuando discrete es TRUE o FALSE}
}
\description{
Constructor de escala de colores de contorno (color) para la dnmye
}
\examples{
library(ggplot2)
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
geom_point(size = 4) +
scale_color_dnmye()
}
