% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_dnmye.R
\name{scale_fill_dnmye}
\alias{scale_fill_dnmye}
\title{Constructor de escala de colores de contorno (color) para la dnmye}
\usage{
scale_fill_dnmye(
  palette = "cualitativa",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Nombre de la paleta de colores en \code{dnmye_paleta()}. Las opciones son \emph{cualitativa} (por default), \emph{secuencial}, \emph{dicotomica} y \emph{divergente}}

\item{discrete}{TRUE (por default) indica que la variable es discreta}

\item{reverse}{\emph{TRUE} invierte el orden de asignación de colores en la paleta}

\item{...}{Argumentos adicionales para definir en discrete_scale() o
scale_color_gradientn(), respectivamente, cuando discrete es TRUE o FALSE}
}
\description{
Constructor de escala de colores de contorno (color) para la dnmye
}
\examples{
datafile <- system.file("toy_evyth.rds", package = "comunicacion")
toy_evyth <- readRDS(datafile)

library(magrittr)
toy_evyth \%>\%
 dplyr::mutate(region_destino = factor(region_destino,
                                       labels = c("Ciudad de Buenos Aires",
                                                  "Provincia de Buenos Aires - Partidos del GBA",
                                                  "Provincia de Buenos Aires - Resto", "Centro",
                                                  "Litoral",
                                                  "Norte", "Cuyo", "Patagonia"))) \%>\%
 dplyr::group_by(anio, region_destino) \%>\%
 dplyr::summarise(cantidad_personas = sum(pondera)) \%>\%
 ggplot2::ggplot(ggplot2::aes(x = anio,
                              y = cantidad_personas,
                              fill = region_destino)) +
 ggplot2::geom_col(position = ggplot2::position_dodge()) +
 scale_fill_dnmye(palette = "cualitativa", discrete = TRUE)
}
