% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_informe.R
\name{template_informe}
\alias{template_informe}
\title{Function for DNMYE pdf template}
\usage{
template_informe(
  logo = "0",
  front_img = "0",
  back_img = "0",
  img_to_dark = FALSE,
  logo_to_white = FALSE,
  other_css = NULL,
  ...
)
}
\arguments{
\item{logo}{Logo}

\item{front_img}{Front cover image}

\item{back_img}{Back cover image}

\item{img_to_dark}{Darken image}

\item{logo_to_white}{Tranform logo color to white - logo must be SVG}

\item{other_css}{Add an other CSS}

\item{...}{Arguments passed to pagedown::html_paged}
}
\value{
A pagedown template
}
\description{
Function for DNMYE pdf template
}
