% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_evyth.R
\docType{data}
\name{toy_evyth}
\alias{toy_evyth}
\title{Muestra no representativa (de "juguete) de la Encuesta de Viajes y Turismo de los Hogares (EVyTH-DNMyE).}
\format{
Un data frame con 5000 filas y 88 variables
}
\source{
\url{https://datos.yvera.gob.ar/dataset/encuesta-de-viajes-y-turismo-de-los-hogares-evyth-microdatos}
}
\usage{
toy_evyth
}
\description{
Base de datos con información del turismo interno (viajes y viajeros)
}
\keyword{datasets}
