% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unir_paginas.R
\name{unir_paginas}
\alias{unir_paginas}
\title{Función para agregar hojas institucionales a los informes generados con el template de la DNMyE}
\usage{
unir_paginas(input, save_as, topic = "generico")
}
\arguments{
\item{input}{Ruta con el PDF del informe}

\item{save_as}{Nombre y ruta del informe a guardar}

\item{topic}{Tema del informe. Puede ser: "calidad", "competitividad hotelera", "eoh", "sector hotelero", "internacional", "interno", "inversiones", "naturaleza", "naturaleza apn" o "generico"}
}
\description{
Función para agregar hojas institucionales a los informes generados con el template de la DNMyE
}
\examples{
\dontrun{
  unir_paginas(input = "eoh.pdf",
   save_as = "informe_eoh_junio_2022.pdf",
   topic = "eoh"
 )
}
}
