% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comunicacion-package.R
\docType{package}
\name{comunicacion-package}
\alias{comunicacion}
\alias{comunicacion-package}
\title{comunicacion: DNMyE - Comunicación}
\description{
Herramientas para la comunicación de la Dirección Nacional de
    Mercados y Estadística.
}
\author{
\strong{Maintainer}: Pablo Tiscornia \email{ptiscornia@turismo.gob.ar} (\href{https://orcid.org/0000-0002-1565-9313}{ORCID})

}
\keyword{internal}
