% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnmye_colores.R
\name{dnmye_colores}
\alias{dnmye_colores}
\title{Funcion para extraer la paleta institucional de la dnmye como colores hex}
\usage{
dnmye_colores(...)
}
\arguments{
\item{...}{Nombre del color. Las posibilidades son \emph{rojo}, \emph{naranja}, \emph{amarillo},
\emph{verde claro}, \emph{verde_oscuro}, \emph{violeta}, \emph{celeste}, \emph{gris oscuro}, \emph{gris medio},
y \emph{gris claro}.}
}
\value{
Paleta de colores institucionales para la Republica de Argentina
}
\description{
Funcion para extraer la paleta institucional de la dnmye como colores hex
}
\examples{
### Obtener el código hex del color institucional celeste
dnmye_colores()

dnmye_colores("rosa")
}
