% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnmye_colores.R
\name{dnmye_colores}
\alias{dnmye_colores}
\title{Función para extraer colores de la dnmye como colores hex}
\usage{
dnmye_colores(...)
}
\arguments{
\item{...}{Nombre del color}
}
\value{
Paleta de colores institucionales para la República de Argentina
}
\description{
Función para extraer colores de la dnmye como colores hex
}
\examples{
dnmye_colores()

dnmye_colores("rojo")

dnmye_colores(1)
}
