% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbl_decimal.R
\name{lbl_decimal}
\alias{lbl_decimal}
\title{Función para formatear números decimales como texto}
\usage{
lbl_decimal(x, decimales = 2)
}
\arguments{
\item{x}{Número a formatear}

\item{decimales}{Cantidad de decimales expresado como numero entero. El valor por default es 2. que recorta a 2 decimales}
}
\description{
Función para formatear números decimales como texto
}
