% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbl_money.R
\name{lbl_money}
\alias{lbl_money}
\title{Función para formatear números que expresan moneda como texto}
\usage{
lbl_money(x, decimales = 1, escala = 1, sufijo = NULL)
}
\arguments{
\item{x}{Número a formatear}

\item{decimales}{Cantidad de decimales. El valor por default es 1 que no muestra decimales}

\item{escala}{Valor para dividir el número a formatear, por ejemplo, en caso de querer expresar el resultado en miles se pone 1000}

\item{sufijo}{Texto a agregar como sufijo}
}
\description{
Función para formatear números que expresan moneda como texto
}
