% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_evyth.R
\docType{data}
\name{toy_evyth}
\alias{toy_evyth}
\title{Base de juguete de la Encuesta de Viajes y Turismo de los Hogares (EVyTH).}
\format{
Una base de datos con 5000 casos y 88 variables:
}
\source{
\url{http://datos.yvera.gob.ar/dataset/encuesta-de-viajes-y-turismo-de-los-hogares-evyth-microdatos}
}
\usage{
toy_evyth
}
\description{
Un dataset que contiene una muestra no representativa de la EVyTH
}
\keyword{datasets}
